Data transport interfaces name change and it's implementation
