--- app/src/main/java/com/example/stretchy/features/datatransport/ExportSavedData.kt
@@ -1,5 +1,0 @@
-package com.example.stretchy.features.datatransport
-
-interface ExportSavedData {
-    suspend fun saveDataInFile()
-}
