--- app/src/main/java/com/example/stretchy/features/datatransport/ImportSavedData.kt
@@ -1,5 +1,0 @@
-package com.example.stretchy.features.datatransport
-
-interface ImportSavedData {
-    fun addSavedDataToDb()
-}
