+++ app/src/main/java/com/example/stretchy/features/traininglist/di/TrainingListModule.kt
@@ -1,7 +1,7 @@
 package com.example.stretchy.features.traininglist.di
 
-import com.example.stretchy.features.datatransport.DataExporter
-import com.example.stretchy.features.datatransport.DataImporter
+import com.example.stretchy.features.datatransport.DataExporterExporterImpl
+import com.example.stretchy.features.datatransport.DataImporterImporterImpl
 import com.example.stretchy.features.traininglist.ui.TrainingListViewModel
 import com.example.stretchy.repository.Repository
 import dagger.Module
@@ -12,8 +12,8 @@ class TrainingListModule {
     @Provides
     fun provideTrainingListViewModel(
         repository: Repository,
-        dataImporter: DataImporter,
-        dataExporter: DataExporter
+        dataImporterImpl: DataImporterImporterImpl,
+        dataExporterImpl: DataExporterExporterImpl
     ) =
-        TrainingListViewModel(repository, dataImporter, dataExporter)
+        TrainingListViewModel(repository, dataImporterImpl, dataExporterImpl)
 }
